# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trackventory API",
        "description": "API documentation for Trackventory - Inventory Tracking System",
        "contact": {
            "name": "Trackventory Support",
            "email": "support@trackventory.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/purchase-invoices": {
            "get": {
                "tags": [
                    "Purchase Invoices"
                ],
                "summary": "Get list of purchase invoices",
                "description": "Retrieve all purchase invoices with search, sort, and pagination",
                "operationId": "getPurchaseInvoicesList",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for item purchased",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "item_purchased",
                                "total_price",
                                "budget",
                                "quantity",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Purchase invoices retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Purchase invoices retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PurchaseInvoice"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "from": {
                                                    "type": "integer"
                                                },
                                                "to": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Purchase Invoices"
                ],
                "summary": "Create a new purchase invoice",
                "description": "Store a newly created purchase invoice",
                "operationId": "storePurchaseInvoice",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PurchaseInvoiceStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Purchase invoice created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Purchase invoice created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PurchaseInvoice"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/purchase-invoices/{id}": {
            "get": {
                "tags": [
                    "Purchase Invoices"
                ],
                "summary": "Get purchase invoice information",
                "description": "Retrieve a specific purchase invoice by ID",
                "operationId": "getPurchaseInvoice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Purchase invoice ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Purchase invoice retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Purchase invoice retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PurchaseInvoice"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Purchase invoice not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Purchase Invoices"
                ],
                "summary": "Update purchase invoice",
                "description": "Update the specified purchase invoice",
                "operationId": "updatePurchaseInvoice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Purchase invoice ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PurchaseInvoiceUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Purchase invoice updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Purchase invoice updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PurchaseInvoice"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Purchase invoice not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Purchase Invoices"
                ],
                "summary": "Delete purchase invoice",
                "description": "Remove the specified purchase invoice",
                "operationId": "deletePurchaseInvoice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Purchase invoice ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Purchase invoice deleted successfully"
                    },
                    "404": {
                        "description": "Purchase invoice not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PurchaseInvoice": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "item_purchased": {
                        "type": "string",
                        "example": "Office Chair"
                    },
                    "total_price": {
                        "type": "number",
                        "format": "float",
                        "example": 299.99
                    },
                    "budget": {
                        "type": "number",
                        "format": "float",
                        "example": 350
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 5
                    },
                    "img": {
                        "type": "string",
                        "example": "https://example.com/image.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "name": [
                                "The name field is required."
                            ],
                            "email": [
                                "The email field must be a valid email address."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "PurchaseInvoiceStoreRequest": {
                "required": [
                    "item_purchased",
                    "total_price",
                    "budget",
                    "quantity"
                ],
                "properties": {
                    "item_purchased": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Office Laptop"
                    },
                    "total_price": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 1200
                    },
                    "budget": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 1500
                    },
                    "quantity": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 3
                    },
                    "img": {
                        "type": "string",
                        "example": "laptop-image.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PurchaseInvoiceUpdateRequest": {
                "properties": {
                    "item_purchased": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Office Laptop"
                    },
                    "total_price": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 1200
                    },
                    "budget": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 1500
                    },
                    "quantity": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 3
                    },
                    "img": {
                        "type": "string",
                        "example": "laptop-image.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}